% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.formula.R
\encoding{UTF-8}
\name{check.formula}
\alias{check.formula}
\title{Internal function}
\usage{
check.formula(formula, vectornames)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}

\item{vectornames}{A vector with names to check the index.}
}
\value{
The index of left side of the formula in vectornames.
}
\description{
Internal function to check the validity of left hand side in a formula object.
}
