% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.names.matrix.p.null.R
\encoding{UTF-8}
\name{mutate.names.matrix.p.null}
\alias{mutate.names.matrix.p.null}
\title{Internal function}
\usage{
mutate.names.matrix.p.null(x, replacement, newname)
}
\arguments{
\item{x}{An object returned by \code{\link{matrix.p.null}}.}

\item{replacement}{A replacement name to matched in object returned by \code{\link{matrix.p.null}}.}

\item{newname}{New name to be replaced in object returned by \code{\link{matrix.p.null}}.}
}
\description{
Internal function to perform replacement names in object returned by \code{\link{matrix.p.null}}.
}
