% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.pcpsmethod.R
\encoding{UTF-8}
\name{select.pcpsmethod}
\alias{select.pcpsmethod}
\title{Internal function}
\usage{
select.pcpsmethod(
  method = c("mantel", "adonis", "glm", "rda", "gls.marginal", "gls.sequential",
    "lme.marginal", "lme.sequential", "none")
)
}
\arguments{
\item{method}{A predefined method/function available in PCPS package, partial match to "mantel", "adonis", "glm", "rda", "gls.marginal", "gls.sequential", "lme.marginal", "lme.sequential" and "none".}
}
\description{
Internal function to select a predefined method/function available in this package.
}
